tests/requirements.txt
pytest
requests
progressbar2
